Putting my emacs.d under version control.
It seems likely I'll be using Emacs as my main editor for quite a long time
at this point (wasting time on the hunt for The Perfect Editor is dangerous,
and anyway, Emacs works just about everywhere), so I'm sucking it up and
starting to really work on getting a decent environment for Emacsing.

I guess we'll see how it goes.