POC dynamic ac-source selection for web-mode
Since this is a proof-of-concept, be wary that it may bite you, and has
not been thoroughly tested.

Note that I hacked my copy of emmet-mode a bit, to verify that adding
the pre-auto-complete hook to web-mode let me work around emmet-mode's
assumption that if we're not in a css-mode variant, we must want HTML
mode.

I think the change to emmet-mode is a good idea in its own right, so
I've sent a pull request to emmet's author.

I did the same thing with php-auto-yasnippet, but it's a small change,
and the merge request is awaiting review.