Copy aggressive-fill-paragraph to site-lisp for hacking
I've made some changes to it, which I don't want to commit in elpa/, as
then they'd be blown away by upgrading.

I'm hoping to contribute them back upstream, but at the moment they're
still evolving, and I want to get them committed to lower the risk of
losing them.