Add lsp-mode from MELPA
There's another one out there, but this is the one in MELPA, so it's the
one I'm going to try for now.

My main goal is just to get better PHP intelligence in emacs, as there
seem to be several LSP backends for PHP.

I'm hopeful that in the long term, LSP will solve many of my 'make Emacs
smart about language X' problems, as that's its promise.

Since it has the might of Microsoft behind it, that doesn't actually
seem impossible.

I guess we'll see how this pans out.