Document a bizarre bug in js2-mode that's been annoying me for years
I had not understood that it had to do with how quickly I typed, which
is really pretty interesting.

I guess it must have to do with how slowly the parser in js2-mode runs.
I remember Steve Yegge saying "I hacked it", and something about relying
on the fact that his parser was pretty quick.

Maybe this is a manifestation of that.