Install atomic-chrome plugin from MELPA and start server at Emacs start
Install the corresponding Chrome plugin and you can edit text in Emacs
at the stroke of a shortcut.

That sounds like it's barely better than copy/pasting, but it hooks the
two up for instantaneous feedback - changes in one are seen almost
immediately in the other, which is gold for sites with live previews.

Also it saves you the pain of having to create and kill temp buffers.