Move magit-svn out of MELPA directory
because it no longer works with the latest version of transient.el and
has itself been unmaintained for years.

Ideally you'd update magit-svn to use the new transient.el release, but
this should keep it working for the foreseeable future.

This is why you keep your dependencies in version control, kids.