Disable auto-filling in atomic-chrome-edit-mode
Since I'm typing into websites, which handle text wrapping themselves, I
don't want to automatically fill in this case.

I'm increasingly nebulous on the value of auto-fill myself. It's great
in source code comments, but I'm not sure it's really a great idea in
plain text files that are just about writing...

dunno.

Anyway, this makes life a little better and wasn't hard to do.