Replace helm with ivy and counsel
Initial results are promising.

We'll see how it does.

This was brought on by a combination of helm suddenly becoming
unmaintained for a while without warning, and a vague sense that
probably ivy/counsel would be more in line with how I like software to
be anyway.

I guess I'll find out.