Move svn-msg-mode auto-mode declaration to config-packages.el
I believe this should get svn-msg-mode working automatically after Emacs
startup. I've had issues with it not doing so on and off, and I think
this is why.

(I mostly don't notice because I usually use git-svn and magit-svn for
working with Subversion repos, but this is relevant when making svn tags
or working directly in an svn checkout for some reason)